SimpleLattices

Documentation for SimpleLattices.

SimpleLattices.CubicLattice

SimpleLattices.SquareLattice

SimpleLattices.plot_lattice

SimpleLattices.CubicLattice — Method

CubicLattice(Nx::Int, Ny::Int, Nz::Int; xperiodic::Bool=false, yperiodic::Bool=false, zperiodic::Bool=false)

Create a cubic lattice with Nx sites in the x-direction, Ny sites in the y-direction, and Nz sites in the z-direction. By default, the lattice is not periodic in any direction. Set xperiodic=true, yperiodic=true, or zperiodic=true to make it periodic in the respective direction.

SimpleLattices.SquareLattice — Method

SquareLattice(Nx::Int, Ny::Int; xperiodic::Bool=false, yperiodic::Bool=false)

Create a square lattice with Nx sites in the x-direction and Ny sites in the y-direction. By default, the lattice is not periodic in either direction. Set xperiodic=true or yperiodic=true to make it periodic in the respective direction.

SimpleLattices.plot_lattice — Method

plot_lattice(lattice::AbstractLattice)

Plot the given lattice. Using GLMakie or CairoMakie if available. Returns a figure.